+++ app/src/main/java/com/univoice/feature/example/compose/ExampleSideEffect.kt
@@ -1,0 +1,5 @@
+package com.univoice.feature.example.compose
+
+sealed class ExampleSideEffect {
+    data class ExampleShowToast(val message: String) : ExampleSideEffect()
+}
