+++ app/src/main/java/com/univoice/feature/example/compose/ExampleState.kt
@@ -1,4 +1,4 @@
-package com.univoice.feature.example_compose
+package com.univoice.feature.example.compose
 
 import com.univoice.core_ui.view.UiState
 import com.univoice.domain.entity.UserEntity
