--- app/src/main/java/com/univoice/feature/example_compose/ExampleSideEffect.kt
@@ -1,5 +1,0 @@
-package com.univoice.feature.example_compose
-
-sealed class ExampleSideEffect {
-    data class ShowToast(val message: String) : ExampleSideEffect()
-}
